Update README.md
Added information about the cipher, we should include a series of starter operations for the user to play with and decode. Also, we should establish if there is a way to have the user enter a string of operations and our language would be able to parse and return a full string that they could then input into our language and decipher.